+++ src/bika/lims/content/rejectanalysis.py
@@ -19,7 +19,8 @@
 # Some rights reserved, see README and LICENSE.
 
 from AccessControl import ClassSecurityInfo
-from Products.Archetypes.public import ReferenceField, Schema, registerType
+from bika.lims.browser.fields import UIDReferenceField
+from Products.Archetypes.public import Schema, registerType
 from bika.lims.content.analysis import Analysis
 from bika.lims.config import PROJECTNAME
 from bika.lims.content.analysis import schema as analysis_schema
@@ -28,10 +29,9 @@ from zope.interface import implements
 
 schema = analysis_schema + Schema((
     # The analysis that was originally rejected
-    ReferenceField(
+    UIDReferenceField(
         'Analysis',
         allowed_types=('Analysis',),
-        relationship='RejectAnalysisAnalysis',
     ),
 ))